# src/frontend/openapi/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "ResourcePulse.Hosting",
    "version": "1.0"
  },
  "paths": {
...
    "/api/project-nodes": {
      "get": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_GetAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoadResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_Create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectNodeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectNodeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectNodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/project-nodes/{id}": {
      "get": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectNodeReadDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectNodeReadDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectNodeReadDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectNodeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectNodeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectNodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/project-nodes/{id}/subtree": {
      "get": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_GetSubtree",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectNodeReadDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectNodeReadDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectNodeReadDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/project-nodes/{id}/reparent": {
      "post": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_Reparent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReparentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReparentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReparentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/project-nodes/{id}/baseline": {
      "post": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_Baseline",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaselineDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BaselineDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BaselineDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/project-nodes/{id}/rebaseline": {
      "post": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_Rebaseline",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RebaselineDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RebaselineDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RebaselineDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/project-nodes/{id}/replan": {
      "post": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_Replan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplanDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplanDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReplanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/project-nodes/{id}/backfill-actuals": {
      "post": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_BackfillActuals",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackfillActualsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BackfillActualsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BackfillActualsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/project-nodes/{id}/recalculate-planned-from-children": {
      "post": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_RecalculatePlannedFromChildren",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/project-nodes/{id}/recalculate-baseline-from-children": {
      "post": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_RecalculateBaselineFromChildren",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/project-nodes/{id}/planning-mode": {
      "post": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_SetPlanningMode",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPlanningModeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPlanningModeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetPlanningModeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/project-nodes/{id}/estimated-work": {
      "put": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_UpdateEstimatedWork",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEstimatedWorkDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEstimatedWorkDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEstimatedWorkDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/project-nodes/{id}/tags": {
      "post": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_AddTag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProjectNodeTagDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProjectNodeTagDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddProjectNodeTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/project-nodes/{id}/tags/{tagId}": {
      "delete": {
        "tags": [
          "ProjectNodes"
        ],
        "operationId": "ProjectNodes_RemoveTag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tagId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/projects/{id}": {
      "put": {
        "tags": [
          "Projects"
        ],
        "operationId": "Projects_UpdateProject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddProjectNodeTagDto": {
        "type": "object",
        "properties": {
          "tagId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "BackfillActualsDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "end": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BaselineDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date"
          },
          "end": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
...
      "CommitmentLevel": {
        "enum": [
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32",
        "x-enum-varnames": [
          "Exploratory",
          "Planned",
          "Committed",
          "Critical"
        ]
      },
...
      "CreateProjectNodeDto": {
        "type": "object",
        "properties": {
          "parentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "nodeType": {
            "$ref": "#/components/schemas/ProjectNodeType"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/ProjectType"
          },
          "commitmentLevel": {
            "$ref": "#/components/schemas/CommitmentLevel"
          },
          "leadResourceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "client": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DerivedExecutionStatus": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32",
        "x-enum-varnames": [
          "NotStarted",
          "InProgress",
          "Completed"
        ]
      },
...
      "HttpValidationProblemDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ProblemDetails"
          },
          {
            "type": "object",
            "properties": {
              "errors": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "nullable": true
              }
            },
            "additionalProperties": { }
          }
        ]
      },
...
      "LoadResult": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { },
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32",
            "default": -1
          },
          "groupCount": {
            "type": "integer",
            "format": "int32",
            "default": -1
          },
          "summary": {
            "type": "array",
            "items": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PlanningMode": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32",
        "x-enum-varnames": [
          "Unspecified",
          "FixedWork",
          "FixedDuration"
        ]
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ProjectNodeReadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "nodeType": {
            "$ref": "#/components/schemas/ProjectNodeType"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "depth": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/ProjectType"
          },
          "commitmentLevel": {
            "$ref": "#/components/schemas/CommitmentLevel"
          },
          "leadResourceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "leadResourceName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "client": {
            "type": "string",
            "nullable": true
          },
          "baselineStart": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "baselineEnd": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "baselinedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "plannedStart": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "plannedEnd": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "actualStart": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "actualEnd": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "planningMode": {
            "$ref": "#/components/schemas/PlanningMode"
          },
          "estimatedWork": {
            "type": "string",
            "format": "date-span",
            "nullable": true
          },
          "scheduleVarianceStart": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "scheduleVarianceEnd": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "forecastVarianceEnd": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isLate": {
            "type": "boolean",
            "nullable": true
          },
          "derivedStatus": {
            "$ref": "#/components/schemas/DerivedExecutionStatus"
          },
          "isProposed": {
            "type": "boolean",
            "nullable": true
          },
          "skillRequirements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectSkillRequirementDto"
            },
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectNodeTagDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProjectNodeType": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32",
        "x-enum-varnames": [
          "Project",
          "Phase",
          "WorkPackage"
        ]
      },
      "ProjectSkillRequirementDto": {
        "type": "object",
        "properties": {
          "skillId": {
            "type": "string",
            "format": "uuid"
          },
          "minLevel": {
            "$ref": "#/components/schemas/SkillLevel"
          }
        },
        "additionalProperties": false
      },
      "ProjectStatus": {
        "enum": [
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32",
        "x-enum-varnames": [
          "Draft",
          "Active",
          "OnHold",
          "Closed",
          "Cancelled"
        ]
      },
      "ProjectType": {
        "enum": [
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32",
        "x-enum-varnames": [
          "Internal",
          "Customer",
          "Investment",
          "Maintenance"
        ]
      },
...
      "RebaselineDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date"
          },
          "end": {
            "type": "string",
            "format": "date"
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReparentDto": {
        "type": "object",
        "properties": {
          "newParentId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "ReplanDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "end": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SetPlanningModeDto": {
        "type": "object",
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/PlanningMode"
          },
          "estimatedWork": {
            "type": "string",
            "format": "date-span",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateEstimatedWorkDto": {
        "type": "object",
        "properties": {
          "estimatedWork": {
            "type": "string",
            "format": "date-span"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/ProjectType"
          },
          "commitmentLevel": {
            "$ref": "#/components/schemas/CommitmentLevel"
          },
          "leadResourceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "client": {
            "type": "string",
            "nullable": true
          },
          "confirmDemoteHardAllocations": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "UpdateProjectNodeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}